refactor: centralize role helpers
