+++ backend/src/middleware/auth/verifyClassOwnership.js
@@ -1,7 +1,6 @@
 const logger = require('../../utils/logger.js');
 const db = require("../../config/database");
-
-const normalizeRole = (role = "") => role.toLowerCase().replace(/\s+/g, "");
+const { isAdminRole } = require("../../utils/role");
 
 module.exports = async function verifyClassOwnership(req, res, next) {
   const classId = req.params.id || req.params.classId;
@@ -14,9 +13,7 @@ module.exports = async function verifyClassOwnership(req, res, next) {
     if (!cls) return res.status(404).json({ message: "Class not found" });
 
     const roles = req.user.roles || [req.user.role];
-    const isAdmin = roles
-      .map((r) => normalizeRole(r))
-      .some((r) => ["admin", "superadmin"].includes(r));
+    const isAdmin = isAdminRole(roles);
 
     if (cls.instructor_id !== req.user.id && !isAdmin) {
       return res.status(403).json({ message: "Access denied" });
